pyproject.toml
[tool.poetry]
name = "bstat"
version = "0.1.0"
description = ""
authors = ["cmareau-git <charles_mareau@hotmail.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pytest = "^8.3.4"
sqlalchemy = "^2.0.37"
kivy-garden = "^0.1.5"
kivy = "^2.3.1"
kivy-garden-matplotlib = "^0.1.1.dev0"
setuptools = "^75.8.0"
matplotlib = "^3.10.0"
kivymd2 = "^2.0.1"
flake8 = "^7.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
